refactor: moving getImage to it's own file
